+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Write warning to log when `content` of a vote is not null.
 - Changed MSRV to 1.68.2.
 - Write warning to log if actor and actor key have different origins.
+- Reject portable actors without WebFinger address.
 
 ## Fixed